Refactor WriteAckQuerier
