+++ crates/relayer-components/src/chain/traits/components/write_ack_querier.rs
@@ -1,6 +1,5 @@
 use cgp_core::prelude::*;
 
-use crate::chain::traits::types::ibc::HasIbcChainTypes;
 use crate::chain::traits::types::ibc_events::write_ack::HasWriteAckEvent;
 use crate::chain::traits::types::packet::HasIbcPacketTypes;
 use crate::std_prelude::*;
@@ -9,8 +8,6 @@ use crate::std_prelude::*;
 #[async_trait]
 pub trait CanQueryWriteAck<Counterparty>:
     HasWriteAckEvent<Counterparty> + HasIbcPacketTypes<Counterparty> + HasErrorType
-where
-    Counterparty: HasIbcChainTypes<Self>,
 {
     async fn query_write_ack_event(
         &self,
