+++ crates/relayer-cosmos/src/impls/chain/components/mod.rs
@@ -1,2 +1,3 @@
 pub mod query_chain_status;
+pub mod query_write_ack_event;
 pub mod send_messages_as_tx;
