+++ crates/relayer-cosmos/src/methods/packet.rs
@@ -1,13 +1,10 @@
 use alloc::sync::Arc;
 
 use ibc_relayer::chain::handle::ChainHandle;
-use ibc_relayer::chain::requests::{Qualified, QueryUnreceivedPacketsRequest};
-use ibc_relayer::link::packet_events::query_write_ack_events;
-use ibc_relayer::path::PathIdentifiers;
+use ibc_relayer::chain::requests::QueryUnreceivedPacketsRequest;
 use ibc_relayer_types::core::ics04_channel::events::WriteAcknowledgement;
 use ibc_relayer_types::core::ics04_channel::packet::{Packet, PacketMsgType, Sequence};
 use ibc_relayer_types::core::ics24_host::identifier::{ChannelId, PortId};
-use ibc_relayer_types::events::IbcEvent;
 use ibc_relayer_types::Height;
 
 use crate::contexts::chain::CosmosChain;
@@ -174,50 +171,6 @@ pub async fn query_is_packet_received<Chain: ChainHandle>(
             let is_packet_received = unreceived_packet.is_empty();
 
             Ok(is_packet_received)
-        })
-        .await
-}
-
-pub async fn query_write_ack_event<Chain: ChainHandle>(
-    chain: &CosmosChain<Chain>,
-    packet: &Packet,
-) -> Result<Option<WriteAcknowledgement>, Error> {
-    let packet = packet.clone();
-
-    chain
-        .with_blocking_chain_handle(move |chain_handle| {
-            let status = chain_handle
-                .query_application_status()
-                .map_err(BaseError::relayer)?;
-
-            let query_height = Qualified::Equal(status.height);
-
-            let path_ident = PathIdentifiers {
-                port_id: packet.destination_port.clone(),
-                channel_id: packet.destination_channel.clone(),
-                counterparty_port_id: packet.source_port.clone(),
-                counterparty_channel_id: packet.source_channel.clone(),
-            };
-
-            let ibc_events = query_write_ack_events(
-                &chain_handle,
-                &path_ident,
-                &[packet.sequence],
-                query_height,
-            )
-            .map_err(BaseError::relayer)?;
-
-            let write_ack = ibc_events.into_iter().find_map(|event_with_height| {
-                let event = event_with_height.event;
-
-                if let IbcEvent::WriteAcknowledgement(write_ack) = event {
-                    Some(write_ack)
-                } else {
-                    None
-                }
-            });
-
-            Ok(write_ack)
         })
         .await
 }
